+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features & Enhancements
 * New option to truncate files to full size (--trunctosize) in creation/write phase.
 * New option to preallocate file disk space (--preallocfile) in creation/write phase.
+* File deletion (-F) can be used when benchmark path is a file.
 
 ## v1.5.3 (Nov 15, 2020)